+++ code/src/cljs/sixsq/nuvla/ui/apps_application/utils.cljs
@@ -4,7 +4,6 @@
 
 
 (def tab-details 5)
-(def tab-docker 6)
 
 
 ;; Deserialization functions: module->db