+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   - Fix bug where apps selected from datasets caused an error
   - Improved read only view of apps
   - Fixed validation errors for apps
+  - Improve layout of apps tabs and move content of deployment tab to overview 
 
 ## [2.20.0] - 2021-08-04